requirements.txt
streamlit
pandas
streamlit-cropper
Pillow
numpy
easyocr